Implement checking runner release file in register runner_version
