+++ pkg/gh/github.go
@@ -1,9 +1,9 @@
 package gh
 
 import (
-	"context"
 	"errors"
 	"fmt"
+	"io"
 	"net/http"
 	"net/url"
 	"path"
@@ -15,7 +15,6 @@ import (
 	"github.com/google/go-github/v35/github"
 	"github.com/m4ns0ur/httpcache"
 	"github.com/patrickmn/go-cache"
-	"github.com/whywaita/myshoes/pkg/logger"
 	"golang.org/x/oauth2"
 )
 
@@ -138,6 +137,24 @@ func CheckSignature(installationID int64) error {
 	return nil
 }
 
+// ExistRunnerReleases check exist of runner file
+func ExistRunnerReleases(runnerVersion string) error {
+	releasesURL := fmt.Sprintf("https://github.com/actions/runner/releases/tag/%s", runnerVersion)
+	resp, err := http.Get(releasesURL)
+	if err != nil {
+		return fmt.Errorf("failed to GET from %s: %w", releasesURL, ErrNotFound)
+	}
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return fmt.Errorf("failed to read response body: %w", err)
+	}
+
+	if resp.StatusCode != 200 {
+		return fmt.Errorf("response status code is not 200: %s", string(body))
+	}
+	return nil
+}
+
 // ExistGitHubRepository check exist of GitHub repository
 func ExistGitHubRepository(scope, gheDomain string, accessToken string) error {
 	repoURL, err := getRepositoryURL(scope, gheDomain)
@@ -164,80 +181,6 @@ func ExistGitHubRepository(scope, gheDomain string, accessToken string) error {
 	}
 
 	return fmt.Errorf("invalid response code (%d)", resp.StatusCode)
-}
-
-// ExistGitHubRunner check exist registered of GitHub runner
-func ExistGitHubRunner(ctx context.Context, client *github.Client, owner, repo, runnerName string) (*github.Runner, error) {
-	runners, err := ListRunners(ctx, client, owner, repo)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get list of runners: %w", err)
-	}
-
-	return ExistGitHubRunnerWithRunner(runners, runnerName)
-}
-
-// ExistGitHubRunnerWithRunner check exist registered of GitHub runner from a list of runner
-func ExistGitHubRunnerWithRunner(runners []*github.Runner, runnerName string) (*github.Runner, error) {
-	for _, r := range runners {
-		if strings.EqualFold(r.GetName(), runnerName) {
-			return r, nil
-		}
-	}
-
-	return nil, ErrNotFound
-}
-
-// ListRunners get runners that registered repository or org
-func ListRunners(ctx context.Context, client *github.Client, owner, repo string) ([]*github.Runner, error) {
-	if cachedRs, found := responseCache.Get(getCacheKey(owner, repo)); found {
-		return cachedRs.([]*github.Runner), nil
-	}
-
-	var opts = &github.ListOptions{
-		Page:    0,
-		PerPage: 100,
-	}
-
-	var rs []*github.Runner
-	for {
-		logger.Logf(true, "get runners from GitHub, page: %d, now all runners: %d", opts.Page, len(rs))
-		runners, resp, err := listRunners(ctx, client, owner, repo, opts)
-		if err != nil {
-			return nil, fmt.Errorf("failed to list runners: %w", err)
-		}
-		storeRateLimit(getRateLimitKey(owner, repo), resp.Rate)
-
-		rs = append(rs, runners.Runners...)
-		if resp.NextPage == 0 {
-			break
-		}
-		opts.Page = resp.NextPage
-	}
-
-	responseCache.Set(getCacheKey(owner, repo), rs, 1*time.Second)
-	logger.Logf(true, "found %d runners in GitHub", len(rs))
-
-	return rs, nil
-}
-
-func getCacheKey(owner, repo string) string {
-	return fmt.Sprintf("owner-%s-repo-%s", owner, repo)
-}
-
-func listRunners(ctx context.Context, client *github.Client, owner, repo string, opts *github.ListOptions) (*github.Runners, *github.Response, error) {
-	if repo == "" {
-		runners, resp, err := client.Actions.ListOrganizationRunners(ctx, owner, opts)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to list organization runners: %w", err)
-		}
-		return runners, resp, nil
-	}
-
-	runners, resp, err := client.Actions.ListRunners(ctx, owner, repo, opts)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to list repository runners: %w", err)
-	}
-	return runners, resp, nil
 }
 
 func getRepositoryURL(scope, gheDomain string) (string, error) {
